converted contact controller to controllerbase
