+++ src/Submissions/Api/Program.cs
@@ -79,7 +79,8 @@ var workersConfig = builder.Configuration.GetSection("Workers").Get<WorkersOptio
 var hangfireConfig = builder.Configuration.GetSection("Hangfire").Get<HangfireOptions>() ?? new();
 
 builder.Services.AddOptions()
-    .Configure<IISServerOptions>(opts => opts.AllowSynchronousIO = true)
+                .Configure<IISServerOptions>(opts => opts.AllowSynchronousIO = true)
+                .Configure<SitePropertiesOptions>(builder.Configuration.GetSection("SiteProperties"))
                 .Configure<JwtBearerConfig>(builder.Configuration.GetSection("JWT"))
                 .Configure<AggregatorOptions>(builder.Configuration.GetSection("Aggregator"))
                 .Configure<AnalyticsOptions>(builder.Configuration.GetSection("Analytics"))
