+++ src/Submissions/Api/Views/Shared/Emails/ExternalNetworkNonMemberInformation.cshtml
@@ -1,4 +1,4 @@
-﻿@inject IOptions<SiteConfigOptions> _config
+﻿@inject IOptions<SitePropertiesOptions> _config
 @inject IHostEnvironment HostEnvironment
 @using Biobanks.Submissions.Api.Config
 @using Biobanks.Submissions.Api.Models.Emails
@@ -18,7 +18,7 @@
 		: $"{directoryName} Contact List (Environment: {HostEnvironment.EnvironmentName})";
 }
 
-<p>
+
 Dear @ViewBag.BiobankName,
 
 Your @ConfigKey.SampleResourceName was selected during a recent search by a user. They have selected to contact you via the @ViewBag.NetworkName system. You do not appear to be part of this system. We believe there is a request waiting for you.
@@ -35,4 +35,3 @@ Please contact us at @contactAddress if you require any further clarification.
 
 Regards,
 @emailSignature
-</p>
