+++ src/Submissions/Api/appsettings.json
@@ -19,5 +19,5 @@
   "DirectoryEnabled": {
     "Enabled": "false"
   },
-    "AllowedHosts": "*"
+  "AllowedHosts": "*"
   }
